move create/deletepack to their own file
